refactor(settings): extract settings components to separate modules
- Move input components (SettingsModalShortInput, SettingsModalLongInput, etc.) to common.tsx
- Extract PresetManager component to its own file
- Extract ThemeController component to its own file
- Extract ImportExportComponent to its own file
- Consolidate settings-related type definitions in new settings.ts file
