+++ src/components/settings/ImportExportComponent.tsx
@@ -1,0 +1,81 @@
+import { useTranslation } from 'react-i18next';
+import { LuEye, LuMessageCircleMore } from 'react-icons/lu';
+import { TbDatabaseExport, TbDatabaseImport } from 'react-icons/tb';
+import { DelimeterComponent, SettingsSectionLabel } from '.';
+import { useAppContext } from '../../context/app';
+import { normalizeUrl } from '../../utils';
+import { downloadAsFile } from '../common';
+
+export function ImportExportComponent({ onClose }: { onClose: () => void }) {
+  const { t } = useTranslation();
+  const { importDB, exportDB } = useAppContext();
+
+  const onExport = async () => {
+    return exportDB().then((data) =>
+      downloadAsFile([JSON.stringify(data, null, 2)], 'llama-ui-database.json')
+    );
+  };
+
+  const onImport = async (e: React.ChangeEvent<HTMLInputElement>) => {
+    const files = e.target.files;
+    if (!files || files.length != 1) return false;
+    const data = await files[0].text();
+    await importDB(data);
+    onClose();
+  };
+
+  const debugImportDemoConv = async () => {
+    const res = await fetch(
+      normalizeUrl('/demo-conversation.json', import.meta.env.BASE_URL)
+    );
+    if (!res.ok) throw new Error(`HTTP error! status: ${res.status}`);
+    const data = await res.text();
+    await importDB(data);
+    onClose();
+  };
+
+  return (
+    <>
+      <SettingsSectionLabel>
+        <LuMessageCircleMore className="lucide w-4 h-4 mr-1 inline" />
+        {t('settings.importExport.chatsSectionTitle')}
+      </SettingsSectionLabel>
+
+      <div className="grid grid-cols-[repeat(2,max-content)] gap-2">
+        <button className="btn" onClick={onExport}>
+          <TbDatabaseExport className="lucide w-4 h-4 mr-1 inline" />
+          {t('settings.importExport.exportBtnLabel')}
+        </button>
+
+        <input
+          id="file-import"
+          type="file"
+          accept=".json"
+          onInput={onImport}
+          hidden
+        />
+        <label
+          htmlFor="file-import"
+          className="btn"
+          aria-label={t('settings.importExport.importBtnLabel')}
+          tabIndex={0}
+          role="button"
+        >
+          <TbDatabaseImport className="lucide w-4 h-4 mr-1 inline" />
+          {t('settings.importExport.importBtnLabel')}
+        </label>
+      </div>
+
+      <DelimeterComponent />
+
+      <SettingsSectionLabel>
+        <LuEye className="lucide w-4 h-4 mr-1 inline" />
+        {t('settings.importExport.technicalDemoSectionTitle')}
+      </SettingsSectionLabel>
+
+      <button className="btn" onClick={debugImportDemoConv}>
+        {t('settings.importExport.importDemoConversationBtnLabel')}
+      </button>
+    </>
+  );
+}
