+++ src/components/settings/index.tsx
@@ -1,0 +1,4 @@
+export * from './common';
+export * from './ImportExportComponent';
+export * from './PresetManager';
+export * from './ThemeController';
